+++ NEWS.md
@@ -1,3 +1,15 @@
+# ropenaq 0.2.0
+
+* The httr dependency has been replaced with crul.
+
+* Now if not giving a value to limit all pages corresponding to a query are automatically retrieved, using async requests.
+
+* Better use of the status page.
+
+# ropenaq 0.1.4
+
+* Thanks to Mikayla Murphy `ropenaq` functions now use the new maximal limit per call which is 10,000.
+
 # ropenaq 0.1.3
 
 * Now all functions outputs a single data.frame with meta and timestamp as attributes. It should be easier to deal with compared to the former format (a list of 3 data.frames).